move lookupApp/AssetResources tests to separate file
